pyproject.toml
[project]
name = "oiv2"
version = "0.2.0"
description = "Compact AI assistant with computer automation - grid-based screen interaction for small LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy-ai>=2.6.27",
    "pydantic>=2.11.1",
]

[project.scripts]
oiv2 = "oiv2.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
